feat: improve summary group info

+++ src/main/kotlin/pubg/services/pubgapi/dataclasses/matches/MatchIncludedStatsData.kt
@@ -1,6 +1,7 @@
 package pubg.services.pubgapi.dataclasses.matches
 
 data class MatchIncludedStatsData(
+    // participant's data
     var DBNOs: String? = null,
     var assists: String? = null,
     var damageDealt: Float? = null,
@@ -9,8 +10,9 @@ data class MatchIncludedStatsData(
     var kills: Int? = null,
     var playerId: String? = null,
     var winPlace: Int? = null,
-    var walkDistance: Float? = null,
+    var walkDistance: Double? = null,
 
+    // roster's data
     var rank: Int? = null,
     var teamId: Int? = null
 )
